Master Thesis - Experiment II
Experiment II was performed on the Frames dataset — a multi-label collection of human-human dialogues where one of two participants is acting as a Wizard-of-Oz assistant. The goal of the experiment was to anticipate the last or middle human response using large language models, taking the histories of the dialogues into account. For this purpose, firstly, the data preparation was performed; secondly, the dialogues were optimally divided into a training and a test dataset; next, a human response generator was developed using LLMs; and finally, since we could not find a suitable classifier for dialogue acts used by the authors of the dataset, we developed and trained our classifier for 20 dialogue acts, to classify the generated utterances.